some teaks on MessageSidebar component three
