+++ src/components/Apps/FBMessenger/MessagesSidebar/MessagesList.tsx
@@ -1,0 +1,15 @@
+import { FunctionComponent  } from "react"
+import styled from 'styled-components'
+
+const MessagesList: FunctionComponent = () => {
+  return (
+    <Wrapper>
+      Message
+    </Wrapper>
+  )
+}
+
+const Wrapper = styled.div`
+`
+
+export default MessagesList
